+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 0.x.x
+
+* [Changed][list] improve performance by shouldUpdateComponent check [#362](https://github.com/demokratie-live/democracy-client/issues/362)
+
 ### 0.7.10
 
 * Search